Added evaluation files.

+++ tests/blameanalysis/gsl/roots-thresholds/roots_B4_TUNED/demo_fn.c
@@ -1,41 +1,41 @@
 #include "demo_fn.h"
 
-double
-quadratic (double x, void *params)
+float
+quadratic (float x, void *params)
 {
   struct quadratic_params *p 
     = (struct quadratic_params *) params;
 
-  double a = p->a;
-  double b = p->b;
-  double c = p->c;
+  float a = p->a;
+  float b = p->b;
+  float c = p->c;
 
   return (a * x + b) * x + c;
 }
 
-double
-quadratic_deriv (double x, void *params)
+float
+quadratic_deriv (float x, void *params)
 {
   struct quadratic_params *p 
     = (struct quadratic_params *) params;
 
-  double a = p->a;
-  double b = p->b;
-  double c = p->c;
+  float a = p->a;
+  float b = p->b;
+  float c = p->c;
 
   return 2.0 * a * x + b;
 }
 
 void
-quadratic_fdf (double x, void *params, 
-               double *y, double *dy)
+quadratic_fdf (float x, void *params, 
+               float *y, float *dy)
 {
   struct quadratic_params *p 
     = (struct quadratic_params *) params;
 
-  double a = p->a;
-  double b = p->b;
-  double c = p->c;
+  float a = p->a;
+  float b = p->b;
+  float c = p->c;
 
   *y = (a * x + b) * x + c;
   *dy = 2.0 * a * x + b;
