+++ tests/blameanalysis/gsl/roots-thresholds/roots_B4_TUNED/demo_fn.h
@@ -2,11 +2,11 @@
 #define LOG_H_DEMO_FN
 
 struct quadratic_params {
-	double a, b, c;
+	float a, b, c;
 };
 
-double quadratic(double x, void* params);
-double quadratic_deriv(double x, void* params);
-void quadratic_fdf(double x, void* params, double* y, double* dy);
+float quadratic(float x, void* params);
+float quadratic_deriv(float x, void* params);
+void quadratic_fdf(float x, void* params, float* y, float* dy);
 
 #endif
